Release delta flowgraph if loadSegment failed
